Changes to be committed: 	modified:   app/Http/Controllers/CuentaController.php 	modified:   app/Http/Controllers/HomeController.php 	modified:   app/Http/Controllers/LoginController.php 	modified:   app/Http/Kernel.php 	modified:   app/Http/Middleware/CheckVista.php 	modified:   app/Models/User.php 	modified:   database/migrations/2025_02_27_173940_agregar_columnas_user.php 	deleted:    resources/views/cuenta/edit.blade.php 	modified:   resources/views/header/base-views.blade.php 	modified:   resources/views/views/usuarios/fichausuarios.blade.php 	modified:   resources/views/views/usuarios/formulariousuario.blade.php 	modified:   routes/web.php
descripcion:

-Se arreglo la funcion para ocultar las vistas a los usuarios que son registrados
-Se agrego la funcion para modificar y eliminar los usuarios registrados.

+++ database/migrations/2025_02_27_173940_agregar_columnas_user.php
@@ -12,11 +12,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->boolean('Usuarios')->default(false)->nullable();
-            $table->boolean('Beneficiarios')->default(false)->nullable();
-            $table->boolean('Especialistas')->default(false)->nullable();
-            $table->boolean('Talleres')->default(false)->nullable();
-            $table->boolean('Asistencia')->default(false)->nullable();
+            $table->boolean('usuarios')->default(false)->nullable();
+            $table->boolean('beneficiarios')->default(false)->nullable();
+            $table->boolean('especialistas')->default(false)->nullable();
+            $table->boolean('talleres')->default(false)->nullable();
+            $table->boolean('asistencias')->default(false)->nullable();
         });
     }
 
@@ -26,11 +26,22 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('Usuarios');
-            $table->dropColumn('Beneficiarios');
-            $table->dropColumn('Especialistas');
-            $table->dropColumn('Talleres');
-            $table->dropColumn('Asistencia');
+            if (Schema::hasColumn('users', 'Usuarios')) {
+                $table->dropColumn('Usuarios');
+            }
+            if (Schema::hasColumn('users', 'Beneficiarios')) {
+                $table->dropColumn('Beneficiarios');
+            }
+            if (Schema::hasColumn('users', 'Especialistas')) {
+                $table->dropColumn('Especialistas');
+            }
+            if (Schema::hasColumn('users', 'Talleres')) {
+                $table->dropColumn('Talleres');
+            }
+            if (Schema::hasColumn('users', 'Asistencia')) {
+                $table->dropColumn('Asistencia');
+            }
         });
     }
 };
+
